lessons/research_agent_part_3/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "The Nova MCP server"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "rich>=14.0.0",
    "opik>=1.8.42",
    "langgraph>=0.5.3",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langchain-perplexity>=0.1.1",
    "langchain-mcp-adapters>=0.1.9",
    "pydantic-settings>=2.10.0",
    "fastmcp>=0.9.0",
    "google-genai>=1.20.0",
    "python-dotenv>=1.0.0",
    "firecrawl>=4.3.6",
    "gitingest>=0.1.5",
    "httpx>=0.28.1",
    "mcp>=1.11.0",
    "nbformat>=5.10.4",
    "psutil>=7.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.0",
    # Cloud SQL connector for GCP deployment
    "cloud-sql-python-connector[asyncpg]>=1.15.0",
]